Improve code quality in different classes
The Storage, Ui, and Siri classes contained nested logic, long
methods, repeated string concatenations, not prominent happy
paths, and multiple levels of abstraction within specifc code
fragments which reduced readability.

The code needed to be more slap-friendly with a clear happy path.
This will make the code easier to read.

Let's refactor these classes by using extract methods so that
the code is not too long and it is written to the highest level
of abstraction.